added confirmation pop up for deletion and finish trip

+++ frontend/src/pages_components/edit_trip_form.js
@@ -45,20 +45,22 @@ function EditTripForm(props) {
 
 
     const deleteTrip = () => {
-        axios({
-                method: "delete",
-                url: "http://127.0.0.1:5000/trip/" + props.trip.id,
-                data: {},  // TODO: add userID
-                headers: { "Content-Type": "application/json" },
-            })
-                .then((res) => {
-                    console.log(res);
+        if (window.confirm("Confirm deletion")) {
+            axios({
+                    method: "delete",
+                    url: "http://127.0.0.1:5000/trip/" + props.trip.id,
+                    data: {},  // TODO: add userID
+                    headers: { "Content-Type": "application/json" },
                 })
-                .catch((err) => {
-                    console.log(err);
-                });
-        props.handleClose();
-        history.push("/trips");
+                    .then((res) => {
+                        console.log(res);
+                    })
+                    .catch((err) => {
+                        console.log(err);
+                    });
+            props.handleClose();
+            history.push("/trips");
+        }
     }
 
     useEffect(() => {
